backend/appwrite_function/requirements.txt
openai-agents==0.8.1
qdrant-client>=1.8.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
openai>=1.0.0
